% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_bar_plot.R
\name{aa_bar_plot}
\alias{aa_bar_plot}
\title{aa bar plot}
\usage{
aa_bar_plot(aa_sequence)
}
\arguments{
\item{aa_sequence}{aa sequence to visualize}
}
\value{
aa bar plot
}
\description{
aa bar plot
}
\examples{
aa_bar_plot("MNAGYWMAMN")
}
