% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_plot.R
\name{aa_plot}
\alias{aa_plot}
\title{Making a plot of aa counts}
\usage{
aa_plot(aa_sequence)
}
\arguments{
\item{aa_sequence}{is a string of amino acid letters}
}
\value{
a bar chart showing the number of each amino acid
}
\description{
Making a plot of aa counts
}
\examples{
aa_plot("CTGPCCTRND")
}
