% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_seq_to_prot.R
\name{convert_seq_to_prot}
\alias{convert_seq_to_prot}
\title{Converting sequence to protein}
\usage{
convert_seq_to_prot(codons)
}
\arguments{
\item{codons}{Sequence that you want to convert}
}
\value{
Amino Acids sequence
}
\description{
Converting sequence to protein
}
