% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_seq.R
\name{create_random_seq}
\alias{create_random_seq}
\title{Create a random sequence}
\usage{
create_random_seq(seq_len)
}
\arguments{
\item{seq_len}{Length of the random sequence}
}
\value{
Sequence randomly generated
}
\description{
Create a random sequence
}
