% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codons.R
\name{extract_codons}
\alias{extract_codons}
\title{Extract codons from RNA sequence}
\usage{
extract_codons(rna_seq, start = 1)
}
\arguments{
\item{rna_seq}{the rna sequence to extract codons from}

\item{start}{start reading position}
}
\value{
a vector of codons
}
\description{
Extract codons from RNA sequence
}
\examples{
extract_codons("AGUGCAUG", 1)
}
