% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna.R
\name{generate_dna}
\alias{generate_dna}
\title{Generate a random DNA sequence}
\usage{
generate_dna(sequence_length)
}
\arguments{
\item{sequence_length}{Integer specifying the length of DNA sequence to generate}
}
\value{
A character string containing a random DNA sequence with bases A, T, G, C
}
\description{
Generate a random DNA sequence
}
\examples{
generate_dna(10)
generate_dna(100)
}
