% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codons.R
\name{get_codons}
\alias{get_codons}
\title{Extracts codons to list from string}
\usage{
get_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{nucleotide sequence to extract codons from}

\item{start}{first codon to include}
}
\value{
list containing string of each codon as an entry
}
\description{
Extracts codons to list from string
}
\examples{
get_codons("ATGGTATTGCGT")
[1] "ATG" "GTA" "TTG" "CGT"
}
