% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_codon.R
\name{make_codon}
\alias{make_codon}
\title{Title Creation of a sequence of codons}
\usage{
make_codon(any_rna_sequence, start = 1)
}
\arguments{
\item{start}{indicating RNA strand as input to the function.}
}
\value{
Return character strings representing the codons
}
\description{
Creates different codons based on an RNA sequence.
}
\examples{
my_codon <- make_codon(my_rna)
my_codon
}
\author{
Oriade Latifah Simpson, \email{s172084@dtu.dk}
}
