% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dna.R
\name{make_dna}
\alias{make_dna}
\title{Title Creation of a sequence of DNA nucleotides.}
\usage{
make_dna(sequence_length)
}
\arguments{
\item{Set}{sequence_length indicating the length of the DNA strand.}
}
\value{
Return character strings representing the DNA.
}
\description{
Creates different lengths of nucleotides of a DNA sequence.
}
\examples{
my_dna <- make_dna(5)
my_dna
}
\author{
Oriade Latifah Simpson, \email{s172084@dtu.dk}
}
