% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_polypeptide.R
\name{make_polypeptide}
\alias{make_polypeptide}
\title{The creation of a polypeptide sequence from the codon.}
\usage{
make_polypeptide(codons)
}
\arguments{
\item{codons}{}
}
\value{
Return character strings representing the polypeptide
}
\description{
Creates a polypeptide sequence of amino acids based on the codon.
}
\examples{
polypeptide_sequence <- make_polypeptide(c("AUG", "CAA", "CGA", "UCA"))
polypeptide_sequence
}
\author{
Oriade Latifah Simpson, \email{s172084@dtu.dk}
}
