% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rna.R
\name{make_rna}
\alias{make_rna}
\title{Title Creation of a sequence of RNA nucleotides from DNA nucleotides}
\usage{
make_rna(the_dna_sequence)
}
\arguments{
\item{the_dna_sequence}{indicating DNA strand as input to the function.}
}
\value{
Return character strings representing the RNA
}
\description{
Substitutes the T for a U when converting from a DNA sequence to an RNA sequence.
}
\examples{
my_rna <- make_rna(my_dna)
my_rna
}
\author{
Oriade Latifah Simpson, \email{s172084@dtu.dk}
}
