% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_T_with_U.R
\name{make_RNA_seq}
\alias{make_RNA_seq}
\title{Replace T with U (Transcription)}
\usage{
make_RNA_seq(DNA_seq)
}
\arguments{
\item{DNA_seq}{String that represents DNA sequence}
}
\value{
Return string that represents RNA sequence
}
\description{
Replace T with U (Transcription)
}
