% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_amino_acid_distribution.R
\name{plotting_amino_acid_distribution}
\alias{plotting_amino_acid_distribution}
\title{Plotting amino acids distribution}
\usage{
plotting_amino_acid_distribution(amino_acid_sequence)
}
\arguments{
\item{amino_acid_sequence}{Sequence of a protein}
}
\value{
Plot of amino acids distribution
}
\description{
Plotting amino acids distribution
}
