% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_codons.R
\name{split_codons}
\alias{split_codons}
\title{Split of a Sequence into Codons}
\usage{
split_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{A string representing a nucleotide sequence - e.g. DNA or RNA.}

\item{start}{An integer specifying the start position in the sequence - default is 1, meaning the function will start splitting from first position.}
}
\value{
Returns a vector of codons (length of 3) extracted from the sequence.
}
\description{
Split of a Sequence into Codons
}
\examples{
# Splitting a DNA sequence starting from the default first position
split_codons("ATGCGTACG")

}
