% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcription.R
\name{Transcribe_DNA}
\alias{Transcribe_DNA}
\title{Title
Transcription}
\usage{
Transcribe_DNA(DNA_seq)
}
\arguments{
\item{DNA_seq}{Input DNA that is trancribed}
}
\value{
RNA sequence transcribed from the input DNA
}
\description{
Title
Transcription
}
\examples{
Transcribe_DNA("ATGCCGTA")
}
