% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe_dna_to_arn.R
\name{transcribe_dna_to_arn}
\alias{transcribe_dna_to_arn}
\title{Transcription}
\usage{
transcribe_dna_to_arn(dna)
}
\arguments{
\item{dna}{}
}
\value{
the transcription of the dna
}
\description{
Transcription
}
\examples{
transcribe_dna_to_arn("ATGCTAGTTCGA")
}
