% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe_dna_to_rna.R
\name{transcribe_dna_to_rna}
\alias{transcribe_dna_to_rna}
\title{Transcription}
\usage{
transcribe_dna_to_rna(dna)
}
\arguments{
\item{dna}{input dna sequence to transcribe}
}
\value{
the transcription of the dna
}
\description{
Transcription
}
\examples{
transcribe_dna_to_rna("ATGCTAGTTCGA")
}
