% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeDNAtoRNA.R
\name{transcribeDNAtoRNA}
\alias{transcribeDNAtoRNA}
\title{Transcription}
\usage{
transcribeDNAtoRNA(DNA)
}
\arguments{
\item{DNA}{}
}
\value{
the transcription of the DNA (RNA)
}
\description{
Transcription
}
\examples{
transcribeDNAtoARN(ATCGGTACGAGCT)
}
