% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation}
\usage{
translation(codons)
}
\arguments{
\item{codons}{a vector of codons}
}
\value{
string of corresponding amino acid sequence for codon vector input.
}
\description{
Translation
}
\examples{
translation(c("AUG", "GGG", "UAA"))
[1] "MG_"
}
