% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator.R
\name{translator}
\alias{translator}
\title{Codon Translator}
\usage{
translator(codons)
}
\arguments{
\item{codons}{codons to be concatenated into a string}
}
\value{
A string of aminoacids(Letters eg. Alanine = A)
}
\description{
Codon Translator
}
\examples{
translator(c("UUU","UCU"))
}
