% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeDNAtoRNA.R
\name{transcribe_dna_to_rna}
\alias{transcribe_dna_to_rna}
\title{Transcription}
\usage{
transcribe_dna_to_rna(DNA)
}
\arguments{
\item{DNA}{}
}
\value{
the transcription of the DNA (RNA)
}
\description{
Transcription
}
\examples{
transcribeDNAtoARN(ATCGGTACGAGCT)
}
